+++ Dockerfile
@@ -4,4 +4,4 @@ COPY . /app
 RUN pip install --upgrade pip && \
     pip install pipenv && \
     pipenv install
-CMD ["python", "-m", "pipenv", "run", "kiki"]
+CMD ["python", "-m", "pipenv", "run", "app"]
